+++ packages/at_secondary_server/lib/src/verb/handler/notify_verb_handler.dart
@@ -105,11 +105,11 @@ class NotifyVerbHandler extends AbstractVerbHandler {
 
   Future<void> _handlePolAuthenticatedConnection(
       HashMap<String, String?> verbParams,
-      InboundConnectionMetadata atConnectionMetadata,
+      InboundConnectionMetadata polConnectionMetadata,
       Response response) async {
     logger.info('Storing the notification ${verbParams[AtConstants.atKey]}');
     var atNotificationBuilder = _populateNotificationBuilder(verbParams,
-        fromAtSign: atConnectionMetadata.fromAtSign!);
+        fromAtSign: polConnectionMetadata.fromAtSign!);
     // If messageType is key, atMetadata is set in "_populateNotificationBuilder"
     // When messageType is text, atNotificationBuilder.atMetadata fields are
     // not applicable except "atMetadata.isEncrypted".
@@ -160,11 +160,11 @@ class NotifyVerbHandler extends AbstractVerbHandler {
         atMetadata = await keyStore.getMeta(cachedNotificationKey);
       }
       var metadata = AtMetadataBuilder(
-              newAtMetaData: atNotificationBuilder.atMetaData,
+              atSign: polConnectionMetadata.fromAtSign!,
+              newAtMetaData: atNotificationBuilder.atMetaData!,
               existingMetaData: atMetadata)
           .build();
-      cachedKeyCommitId = await _storeCachedKeys(
-          cachedNotificationKey, metadata,
+      cachedKeyCommitId = await _storeCachedKey(cachedNotificationKey, metadata,
           atValue: atNotificationBuilder.atValue);
       //write the latest commit id to the StatsNotificationService
       _writeStats(cachedKeyCommitId, operationType.name);
@@ -236,7 +236,7 @@ class NotifyVerbHandler extends AbstractVerbHandler {
   /// key Key to cache.
   /// AtMetadata metadata of the key.
   /// atValue value of the key to cache.
-  Future<int> _storeCachedKeys(String? cachedKey, AtMetaData? atMetaData,
+  Future<int> _storeCachedKey(String? cachedKey, AtMetaData? atMetaData,
       {String? atValue}) async {
     var atData = AtData();
     atData.data = atValue;
@@ -381,6 +381,10 @@ class NotifyVerbHandler extends AbstractVerbHandler {
     if (verbParams[AtConstants.sharedKeyEncryptedEncryptingAlgo] != null) {
       atMetadata.skeEncAlgo =
           verbParams[AtConstants.sharedKeyEncryptedEncryptingAlgo];
+    }
+    if (verbParams[AtConstants.immutable] != null) {
+      atMetadata.immutable =
+          SecondaryUtil.getBoolFromString(verbParams[AtConstants.immutable]);
     }
     atMetadata.isEncrypted = getIsEncrypted(
         getMessageType(verbParams[AtConstants.messageType]),