+++ packages/at_secondary_server/test/resource_manager_test.dart
@@ -141,30 +141,34 @@ void main() async {
           'id:1234:messageType:text:notifier:wavi:ttln:900000:@bob:phone@alice');
     });
 
-    test(
-        'Test to verify prepare an update notification command with a value and all the metadata',
-        () {
+    testNotificationMetaData({required bool immutable}) {
       var ttln = 24 * 60 * 60 * 1000;
+      var fromAtsign = '@alice';
       var atNotification = (AtNotificationBuilder()
-            ..fromAtSign = '@alice'
+            ..fromAtSign = fromAtsign
             ..toAtSign = '@bob'
             ..id = '1234'
             ..opType = OperationType.update
             ..messageType = MessageType.key
             ..atValue = 'Hi Bob, Alice here'
-            ..notification = '@bob:test.test@alice'
+            ..notification = '@bob:test.test$fromAtsign'
             ..notificationDateTime = DateTime.now().toUtcMillisecondsPrecision()
             ..ttl = ttln
-            ..atMetaData = AtMetaData.fromCommonsMetadata(Metadata()
-              ..ttr = 1
-              ..ccd = true
-              ..pubKeyCS = '123'
-              ..sharedKeyEnc = 'abc'
-              ..encKeyName = 'ekn'
-              ..encAlgo = 'ea'
-              ..ivNonce = 'ivn'
-              ..skeEncKeyName = 'ske_ekn'
-              ..skeEncAlgo = 'ske_ea'))
+            ..atMetaData = AtMetaData.fromCommonsMetadata(
+              Metadata()
+                ..ttr = 1
+                ..ccd = true
+                ..pubKeyCS = '123'
+                ..sharedKeyEnc = 'abc'
+                ..encKeyName = 'ekn'
+                ..encAlgo = 'ea'
+                ..ivNonce = 'ivn'
+                ..skeEncKeyName = 'ske_ekn'
+                ..skeEncAlgo = 'ske_ea'
+                ..pubKeyHash = PublicKeyHash('someHash', 'someAlgo')
+                ..immutable = immutable,
+              fromAtsign,
+            ))
           .build();
 
       var notifyCommand = ResourceManager.getInstance()
@@ -180,10 +184,19 @@ void main() async {
           ':ttr:1:ccd:true'
           ':isEncrypted:false'
           ':sharedKeyEnc:abc:pubKeyCS:123'
+          ':pubKeyHash:someHash:hashingAlgo:someAlgo'
           ':encKeyName:ekn:encAlgo:ea:ivNonce:ivn'
           ':skeEncKeyName:ske_ekn:skeEncAlgo:ske_ea'
+          '${immutable ? ':immutable:true' : ''}'
           ':@bob:test.test@alice'
           ':Hi Bob, Alice here');
+    }
+
+    test(
+        'Test to verify prepare an update notification command with a value and all the metadata',
+        () {
+      testNotificationMetaData(immutable: false);
+      testNotificationMetaData(immutable: true);
     });
   });
 }