+++ tests/at_functional_test/lib/utils/encryption_util.dart
@@ -1,11 +1,9 @@
-
 import 'dart:typed_data';
 
 import 'package:crypton/crypton.dart';
 import 'package:encrypt/encrypt.dart';
 
 class EncryptionUtil {
-
   static IV getIV(String? ivBase64) {
     if (ivBase64 == null) {
       return IV(Uint8List(16));